requirements.txt
asgiref==3.7.2
asttokens==2.4.1
black==23.12.1
click==8.1.7
decorator==5.1.1
Django==4.2.9
django-extensions==3.2.3
django-filter==23.5
django-phonenumber-field==7.3.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
drf-yasg==1.21.7
executing==2.0.1
flake8==7.0.0
graphviz==0.20.1
inflection==0.5.1
ipython==8.20.0
isort==5.13.2
jedi==0.19.1
Markdown==3.5.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
phonenumbers==8.13.28
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pydot==2.0.0
pyflakes==3.2.0
Pygments==2.17.2
pyparsing==3.1.1
pytz==2023.4
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
uritemplate==4.1.1
wcwidth==0.2.13
